requirements.txt
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=24.1.0
neo4j>=5.15.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
qdrant-client>=1.7.0
sentence-transformers>=2.3.1
openai>=1.10.0
tenacity>=8.2.3
fake-useragent>=1.4.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
unstructured>=0.11.0
duckduckgo-search>=4.0.0
geopy>=2.4.0
sqlalchemy>=2.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
